+++ packages/analytics/dashboard-renderer/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.15.2...@kong-ui-public/dashboard-renderer@0.15.3) (2024-03-21)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
+## [0.15.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.15.1...@kong-ui-public/dashboard-renderer@0.15.2) (2024-03-21)
+
+
+### Bug Fixes
+
+* **dashboard-renderer:** golden signals renderer refresh interval ([#1276](https://github.com/Kong/public-ui-components/issues/1276)) ([08532d4](https://github.com/Kong/public-ui-components/commit/08532d442132083cad8fa497188d4a27923a9ca8))
+
+
+
+
+
+## [0.15.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.15.0...@kong-ui-public/dashboard-renderer@0.15.1) (2024-03-21)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
+# [0.15.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.14.6...@kong-ui-public/dashboard-renderer@0.15.0) (2024-03-20)
+
+
+### Features
+
+* allow dashbaord renderer to optionally auto refresh queries ([#1271](https://github.com/Kong/public-ui-components/issues/1271)) ([eee8e28](https://github.com/Kong/public-ui-components/commit/eee8e28d0333935782b2570be27881be6beced01))
+
+
+
+
+
 ## [0.14.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.14.5...@kong-ui-public/dashboard-renderer@0.14.6) (2024-03-20)